Standardized vJoy Interface axis handling
This change standardizes that all axis updates should be sent to the vJoyInterface as normalized vectors. The conversion then happens inside the vJoyInterface class before setting the iReport values in a much more standardized manner in the Update method.

Also added Humanoid methods to allow axis updates for look/move/item wheel

This change might allow us to remove the secondary map axis in the future...since we won't have to do so much customization when interacting with this class from other classes.