Added OnFoot holographic control panels
Added groups of controls to the left, center, and right panels. These controls did not "fit" on the controller bindings, so they have to be set up with Holographic buttons instead.
Rather than requiring the user to map all of these ControlButtons individually, these panels provide all of the remaining required controls.