Fix map error in EVRC.binds
You can't have overlapping bindings between CamTranslate and UI controls. When the game is in map mode (galaxy or system), it will only execute the camtranslate controls
So, the EVRC controls that allow you to navigate either the UI or the Map itself won't work unless those controls are uniquely bound.