Removed ResetAll
this method was added when I was having a problem with throttle overlapping the Z axis movement in the Galaxy map, which turned out to be a bindings problem (although this fixed it at the time).

Now, it's causing a problem where the joystick goes fully Left and Forward every time you exit the map. Resetting the axis already happens in the Update method of the VJoyInterface, so I'm removing this.